separate strain_statistics plot function from the plotter and put it in a separater more general plot.stats module
